requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpsx==0.0.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.2.0
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
